Clean up syllabus upload related functions
